Fix warnings and inconsistent module layout
When --all-features was used, the ssl transport
was provided via the `transport_ssl` module.
However, the examples were not prepared for this,
and `cargo test --all-features` would fail.

As both transports are mutually exclusive, it
seems reasonable to just use the same module name
and make life easier for clients.