Implement conversions for non-string keyed maps
Currently only maps with keys of type String can be used.
This commit adds a macro to automatically generate the AsRust trait
for many different types of hash maps.
It also replaces the implementation of From<HashMap<String, T>> for
Bytes with the more general form From<HashMap<K, V>>.